refactor: eval workflow を1ファイルに統合
- eval-workflow.ts, eval-batch-workflow.ts, eval-helpers.ts を統合
- 共通処理を runEval 関数に集約
- utils ディレクトリを削除
